Add and test an example yaml
